pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer"
authors = ["BRODER1CK <BRODER1CK@mail.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.2"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.1"
jinja2 = "^3.1.2"
psycopg = "^3.1.8"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
